auth_service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-CORS==4.0.0
Flasgger==0.9.7.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
prometheus-client==0.18.0
prometheus-flask-exporter==0.22.4
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.6.1 